editing for clarity and consistency
I made a number of edits, and added a few questions preceded by -- 

Some of the reasoning behind the edits and questions is summarized below:

The section should explain in a clear and straight-forward language the purpose of the function and how it works.  I have inserted as much as I know, but would like to add more given a better understanding. The content should  avoid circular explanations, e.g. "product qualification is inferred from... qualification logic"

The fonts should be consistent throughout the document, e..g no random use of the bold font.

I suggest limiting the documentation to the CDM current state.  Unless absolutely necessary to explain a current state, the documentation should exclude statements describing what it does not do, which generally does not help to explain or promote the product.  Additionally, I suggest that we don't reference future plans - - that information belongs somewhere else and moreover, future statements give the impression that CDM is not finished.

If we are going to refer the reader to another section, then we should do it at the end of the current section rather then interrupting the thought process of the reader.  We should also make sure that any hyperlinks bring the user directly to the advertised section.

That example of productIdentification does not look like a "serialized" version of a message.  My understanding of a serialized version is one that does not have any attribute names because it's associated with a specific scheme for which the attributes can be inferred.  This example looks like a "de-serialized" version, which is a qualification that is not necessary to explain this function.