Func doc update (#789)
* Update documentation.rst

Revision of the Product Section

* A few adjustments

Normalised updated snippet formats, plus minor text edits - for instance: anglicised where needs be (to keep consistent)

* Hyperlink

Hyperlink the Textual Browser in the ISDA CDM Portal

* Create TradableProduct.rst

* Create TradableProduct.rosetta

* Typo

* Delete TradableProduct.rst

* Create TradableProduct.rosetta

Copied from source/snippets

* Delete TradableProduct.rosetta

Copied over to source/code-snippets

* Change file suffix for snippet to .snippet

* First draft of additional documentation

* update scope

* Move some function bits to DSL doc

Those bits have been slightly reworded and merged with other changes in the Rosetta DSL doc. Also, old part of the CDM function doc has been cut-out, so rst can render in GitHub.

* Port over validation section from DSL doc

* Update product qualification section

Also update some snippet examples to Haskell rendering

* Start editing process model section

* Re-arrange some of Jim's changes

* Re-introduce old function sections

* Also keep utility function section

* Update examples in process model section

* Move bits to code-gen doc

* Small update

* Remove duplicate para

* Fix boo-boo from previous conflict-merge

* Link for one-of condition

* editing for clarity and consistency

I made a number of edits, and added a few questions preceded by -- 

Some of the reasoning behind the edits and questions is summarized below:

The section should explain in a clear and straight-forward language the purpose of the function and how it works.  I have inserted as much as I know, but would like to add more given a better understanding. The content should  avoid circular explanations, e.g. "product qualification is inferred from... qualification logic"

The fonts should be consistent throughout the document, e..g no random use of the bold font.

I suggest limiting the documentation to the CDM current state.  Unless absolutely necessary to explain a current state, the documentation should exclude statements describing what it does not do, which generally does not help to explain or promote the product.  Additionally, I suggest that we don't reference future plans - - that information belongs somewhere else and moreover, future statements give the impression that CDM is not finished.

If we are going to refer the reader to another section, then we should do it at the end of the current section rather then interrupting the thought process of the reader.  We should also make sure that any hyperlinks bring the user directly to the advertised section.

That example of productIdentification does not look like a "serialized" version of a message.  My understanding of a serialized version is one that does not have any attribute names because it's associated with a specific scheme for which the attributes can be inferred.  This example looks like a "de-serialized" version, which is a qualification that is not necessary to explain this function.

* Adjust product qualification section

* Adjust serialised JSON example

* Adjust reference to taxonomy v1.0

* Edits to Product Qual and Process Model

Edits in Product Qual reviewed with Leo (except for a couple of exceptions)
Edits in Process Model done solo
- I don't fully understand the intended meaning of the granularity section.  I made edits to help make it more clear according to what I believe is intended, but was not sure it's correct or complete
I did not review any other sections

* fixed a few typos in Model Process

* one last typo

* Minor adjustments

Co-authored-by: TSanborn-ISDA <61701717+TSanborn-ISDA@users.noreply.github.com>
Co-authored-by: Jim Wang <jim.wang@regnosys.com>