A number of changes
Use italics rather than quotes to highlight a new term - Style Guide (SG) 6.3.2
Oxford comma SG 6.1.4.2.4 (Otherwise a very nice summary of the credit support documents)
CDM is always preceded with "the" SG 2.1 (this is the first rule)
Describe current state SG 5.1.2
Moved a set of paragraphs with good background explanation from the description of the model to the background description SG 5.2.3
Use 3rd person, not I, we, or you SG 6.1.7
The bullet on ISDA create was focused on the wrong thing.  The main point is that synonyms have been created to allow automated transformation.  The ingestion is a valuable illustration, not the end purpose.
We did not define a style guide rule for not using NB:, but I will add one.  The approach used in the documentation is .. note::
This is some type of grammar problem "is a data type used to specify whether Umbrella Agreement terms are applicable, relevant specific language, and underlying entities associated with the umbrella agreement."
Suggested revision "is a data type used to specify the applicability of Umbrella Agreement terms, relevant specific language, and underlying entities associated with the umbrella agreement.