Update documentation.rst
Made changes to reflect the changes from the PriceQuanty phase 2 initial release. 

Also removed this random statement about the FpML synonyms which had been sandwiched into the explanation of the Payout data type and had not previously been excavated.  "The relationship between one of the payout classes and a similar structure in FpML can be identified through the defined Synonyms, as explained in an earlier section. For example, the InterestRatePayout is equivalent to the following complex types in FpML: swapStream, feeLeg capFloorStream, fra, and interestLeg."

There is a brief note about synonyms at the beginning of the CDM topic and a generic section about synonyms in the Rosetta documentation.  If we want to add a section to elaborate about synonyms for selected models then we should add a specific section for it to summarize the set of mappings for the whole model rather than inserting a paragraph into one of the sections for which there is mapping.