Financial commodity products contribution.
This contribution enhances the current structure for financial commodity swaps. It includes a new schedule structure for both legs, a new fixed price schedule and an update for the quantity schedule. It also includes a new commodity description element, the mapping of the instrumentId and the extension for the quantity frequency mapping.