Financial commodity products (#1570)
* Financial commodity products update

Contribution of a schedule element for commodity swaps, a new fixed price schedule, an update on the quantity schedule, a new commodity description element and expansion on some commodity mappings

* Financial commodity products contribution.

This contribution enhances the current structure for financial commodity swaps. It includes a new schedule structure for both legs, a new fixed price schedule and an update for the quantity schedule. It also includes a new commodity description element, the mapping of the instrumentId and the extension for the quantity frequency mapping.

* Fix mappings, update expectations

* Fix ProductIdentification.productIdentifier mappings

* Release note

* Remove unused imports

Co-authored-by: mcarrera_auth0_619cb6a21f6a43006b4efbd6 <mcarrera@tradeheader.com>
Co-authored-by: hugohills <hugohills@hotmail.com>