Various small fixes (#1575)
* Financial commodity products update

Contribution of a schedule element for commodity swaps, a new fixed price schedule, an update on the quantity schedule, a new commodity description element and expansion on some commodity mappings

* Financial commodity products contribution.

This contribution enhances the current structure for financial commodity swaps. It includes a new schedule structure for both legs, a new fixed price schedule and an update for the quantity schedule. It also includes a new commodity description element, the mapping of the instrumentId and the extension for the quantity frequency mapping.

* Make masterConfirmationType optional

Master Confirmation Type under agreement name should be optional with a cardinality of 0..1. Contracts such as interest rate derivatives do not have master confirmation agreements

* Fix mappings, update expectations

* Fix ProductIdentification.productIdentifier mappings

* Release note

* Remove unused imports

* frequency conditions

frequency conditions

* Update expectations

* Fix snippet and condtion

Co-authored-by: mcarrera_auth0_619cb6a21f6a43006b4efbd6 <mcarrera@tradeheader.com>
Co-authored-by: hugohills <hugohills@hotmail.com>
Co-authored-by: jon.nance_auth0_5c3dfb00755c3e1c64d1b702 <jon.nance@axoni.com>
Co-authored-by: nigel.cobb_auth0_5f3ab5ac9580dd006d0e78be <nigel.cobb@regnosys.com>