Adding party specific eligibility criteria
This addition of an attribute based on Party type in the EligibleCollateralCriteria will allow specify that a criteria applies to only one party or many parties. These criteria are independent of the parties to a collateral document (e.g. a criteria and treatment may only apply to one party) and may be use when CollateralSpecification is used outside documentation such as for exchange of inventory or eligibility schedules.