PTRR - Portfolio Rebalancing
For the EMIR jurisdiction, there are some reportable fields regarding PTRR (Post-Trade Risk Reduction) events. This release introduces support to allow the reporting for one of these PTRR events, specifically for the Portfolio Rebalancing, by adding its corresponding value to the intent enumeration and a specific qualification function for this event.